Removed dest_reg from the ALU, renamed ARM1176JZF_S.hpp/cpp to amr1176jzf_s.hpp/cpp
