+++ test/CMakeLists.txt
@@ -4,7 +4,7 @@ include_directories(../src)
 # List of src files to be compiler with the unit tests
 set(SRC_FILES
     ../src/cpu/registers/cspr.cpp
-    ../src/cpu/ARM1176JZF_S.cpp
+    ../src/cpu/arm1176jzf_s.cpp
     ../src/cpu/alu.cpp
     ../src/cpu/isa/instruction.cpp
     ../src/cpu/isa/data_processing.cpp
